Publish complete checked expression facts and retire runtime re-resolution
Seal value/contextual records with owner, revision, source identity, types,
shapes, operations, constructor applications and static obligations.
Specialize retained instance and bound products, checking new binding roots
independently and rejecting deferred descendants before operand work.

Remove interpreter constructor-generic resolution, raw Nat axis replay and
shape/declaration-type fallbacks. Preserve actual constructor membership,
domain, finite and dynamic-index validation. Complete nominal-bound
constructor preparation and linear contextual-root completion.

Validated with 3080 workspace tests, four explicit Lean oracles, native lint,
13 Node/Wasm tests, report smoke and the unchanged docs Wasm size gate.
Layer debt shrinks from 95 to 90 exceptions.

+++ crates/graphcal-compiler/src/expression_id.rs
@@ -11,7 +11,8 @@ struct Revision;
 #[derive(Debug, Clone)]
 pub struct ExprId {
     revision: Arc<Revision>,
-    ordinal: u64,
+    // A position in an in-memory body, not a language Nat or serialized identity.
+    ordinal: usize,
 }
 
 impl PartialEq for ExprId {
@@ -34,7 +35,7 @@ pub struct UnassignedExprId;
 /// Construction-only allocator. Each allocator represents a fresh body revision.
 pub(crate) struct ExprIds {
     revision: Arc<Revision>,
-    next: u64,
+    next: usize,
 }
 
 impl Default for ExprIds {
@@ -86,9 +87,9 @@ mod tests {
     fn exhausted_allocator_does_not_publish_or_wrap() {
         let mut ids = ExprIds {
             revision: Arc::new(Revision),
-            next: u64::MAX,
+            next: usize::MAX,
         };
         assert!(ids.allocate().is_err());
-        assert_eq!(ids.next, u64::MAX);
+        assert_eq!(ids.next, usize::MAX);
     }
 }
